refactor: extract visual toolbar classes to widgets/toolbar-ui
